requirements.txt
streamlit
pandas
mysql-connector-python
pymysql
SQLAlchemy